tcassistant-api/requirements.txt
beautifulsoup4==4.13.3
faiss_cpu==1.10.0
flasgger==0.9.7.1
Flask==3.1.0
Flask_RESTful==0.3.10
langchain==0.3.21
langchain_community==0.3.20
langchain_core==0.3.48
langchain_ollama==0.3.0
langchain_text_splitters==0.3.7
numpy==2.2.4
Requests==2.32.3
tokenizers==0.21.1
Werkzeug==3.1.3
